Overhauled tenant_controller logic to work with namespace-scoped tenants
